refactored cli
